requirements.txt
langchain_openai
browser_use
python-dotenv